+++ src/transaction-runner.ts
@@ -121,6 +121,7 @@ export abstract class Runner<T> {
     this.attempts = 0;
     this.session = session;
     this.transaction = transaction;
+    this.transaction.useInRunner();
 
     const defaults = {timeout: 3600000};
 
@@ -197,7 +198,9 @@ export abstract class Runner<T> {
       undefined,
       this.options.optimisticLock
     );
-    await transaction.begin();
+    if (this.attempts > 0) {
+      await transaction.begin();
+    }
     return transaction;
   }
   /**