+++ system-test/spanner.ts
@@ -23,7 +23,7 @@ import * as crypto from 'crypto';
 import * as extend from 'extend';
 import * as is from 'is';
 import * as uuid from 'uuid';
-import {Backup, Database, Spanner, Instance} from '../src';
+import {Backup, Database, Spanner, Instance, InstanceConfig} from '../src';
 import {Key} from '../src/table';
 import {
   ReadRequest,
@@ -36,6 +36,7 @@ import {grpc, CallOptions} from 'google-gax';
 import {google} from '../protos/protos';
 import CreateDatabaseMetadata = google.spanner.admin.database.v1.CreateDatabaseMetadata;
 import CreateBackupMetadata = google.spanner.admin.database.v1.CreateBackupMetadata;
+import CreateInstanceConfigMetadata = google.spanner.admin.instance.v1.CreateInstanceConfigMetadata;
 
 const SKIP_BACKUPS = process.env.SKIP_BACKUPS;
 const PREFIX = 'gcloud-tests-';
@@ -83,12 +84,19 @@ describe('Spanner', () => {
     },
     gaxOptions: GAX_OPTIONS,
   };
+
   const IS_EMULATOR_ENABLED =
     typeof process.env.SPANNER_EMULATOR_HOST !== 'undefined';
   const RESOURCES_TO_CLEAN: Array<Instance | Backup | Database> = [];
+  const INSTANCE_CONFIGS_TO_CLEAN: Array<InstanceConfig> = [];
   const DATABASE = instance.database(generateName('database'), {incStep: 1});
   const TABLE_NAME = 'Singers';
   const PG_DATABASE = instance.database(generateName('pg-db'), {incStep: 1});
+
+  // Custom instance configs start with 'custom-'
+  const instanceConfig = spanner.instanceConfig(
+    'custom-' + generateName('instance-config')
+  );
 
   before(async () => {
     await deleteOldTestInstances();
@@ -131,40 +139,72 @@ describe('Spanner', () => {
       );
       await postgreSqlOperationUpdateDDL.promise();
       RESOURCES_TO_CLEAN.push(PG_DATABASE);
+
+      // Create a user-managed instance config from a base instance config.
+      const [baseInstanceConfig] = await spanner.getInstanceConfig(
+        INSTANCE_CONFIG.config
+      );
+      const customInstanceConfigRequest = {
+        replicas: baseInstanceConfig.replicas!.concat(
+          baseInstanceConfig!.optionalReplicas![0]
+        ),
+        baseConfig: baseInstanceConfig.name,
+        gaxOptions: GAX_OPTIONS,
+      };
+      const [, operation] = await instanceConfig.create(
+        customInstanceConfigRequest
+      );
+      await operation.promise();
+      INSTANCE_CONFIGS_TO_CLEAN.push(instanceConfig);
     }
   });
 
   after(async () => {
-    if (generateInstanceForTest) {
-      // Deleting all backups before an instance can be deleted.
-      await Promise.all(
-        RESOURCES_TO_CLEAN.filter(resource => resource instanceof Backup).map(
-          backup => backup.delete(GAX_OPTIONS)
-        )
-      );
-      /**
-       * Deleting instances created during this test.
-       * All databasess will automatically be deleted with instance.
-       * @see {@link https://cloud.google.com/spanner/docs/reference/rpc/google.spanner.admin.instance.v1#google.spanner.admin.instance.v1.InstanceAdmin.DeleteInstance}
-       */
-      await Promise.all(
-        RESOURCES_TO_CLEAN.filter(resource => resource instanceof Instance).map(
-          instance => instance.delete(GAX_OPTIONS)
-        )
-      );
-    } else {
-      /**
-       * Limit the number of concurrent 'Administrative requests per minute'
-       * Not to exceed quota
-       * @see {@link https://cloud.google.com/spanner/quotas#administrative_limits}
-       */
-      const limit = pLimit(5);
-      await Promise.all(
-        RESOURCES_TO_CLEAN.map(resource =>
-          limit(() => resource.delete(GAX_OPTIONS))
-        )
-      );
+    try {
+      if (generateInstanceForTest) {
+        // Sleep for 30 seconds before cleanup, just in case
+        await new Promise(resolve => setTimeout(resolve, 30000));
+        // Deleting all backups before an instance can be deleted.
+        await Promise.all(
+          RESOURCES_TO_CLEAN.filter(resource => resource instanceof Backup).map(
+            backup => backup.delete(GAX_OPTIONS)
+          )
+        );
+        /**
+         * Deleting instances created during this test.
+         * All databasess will automatically be deleted with instance.
+         * @see {@link https://cloud.google.com/spanner/docs/reference/rpc/google.spanner.admin.instance.v1#google.spanner.admin.instance.v1.InstanceAdmin.DeleteInstance}
+         */
+        await Promise.all(
+          RESOURCES_TO_CLEAN.filter(
+            resource => resource instanceof Instance
+          ).map(instance => instance.delete(GAX_OPTIONS))
+        );
+      } else {
+        /**
+         * Limit the number of concurrent 'Administrative requests per minute'
+         * Not to exceed quota
+         * @see {@link https://cloud.google.com/spanner/quotas#administrative_limits}
+         */
+        const limit = pLimit(5);
+        await Promise.all(
+          RESOURCES_TO_CLEAN.map(resource =>
+            limit(() => resource.delete(GAX_OPTIONS))
+          )
+        );
+      }
+    } catch (err) {
+      console.error('Cleanup failed:', err);
     }
+    /**
+     * Deleting instance configs created during this test.
+     * @see {@link https://cloud.google.com/spanner/docs/reference/rpc/google.spanner.admin.instance.v1#google.spanner.admin.instance.v1.InstanceAdmin.DeleteInstanceConfig}
+     */
+    await Promise.all(
+      INSTANCE_CONFIGS_TO_CLEAN.map(instanceConfig =>
+        instanceConfig.delete({gaxOpts: GAX_OPTIONS})
+      )
+    );
   });
 
   describe('types', () => {
@@ -282,7 +322,8 @@ describe('Spanner', () => {
                   "StringValue"     VARCHAR,
                   "TimestampValue"  TIMESTAMPTZ,
                   "DateValue"       DATE,
-                  "CommitTimestamp" SPANNER.COMMIT_TIMESTAMP
+                  "CommitTimestamp" SPANNER.COMMIT_TIMESTAMP,
+                  "JsonbValue"      JSONB
                 );
             `
         );
@@ -1184,6 +1225,37 @@ describe('Spanner', () => {
       });
     });
 
+    describe('jsonb', () => {
+      before(async function () {
+        if (IS_EMULATOR_ENABLED) {
+          this.skip();
+        }
+      });
+
+      const jsonbInsert = (done, dialect, value) => {
+        insert({JsonbValue: value}, dialect, (err, row) => {
+          assert.ifError(err);
+          assert.deepStrictEqual(row.toJSON().JsonbValue, value);
+          done();
+        });
+      };
+
+      it('POSTGRESQL should write jsonb values', done => {
+        jsonbInsert(
+          done,
+          Spanner.POSTGRESQL,
+          Spanner.pgJsonb({
+            key1: 'value1',
+            key2: 'value2',
+          })
+        );
+      });
+
+      it('POSTGRESQL should write null jsonb values', done => {
+        jsonbInsert(done, Spanner.POSTGRESQL, null);
+      });
+    });
+
     describe('commit timestamp', () => {
       const commitTimestamp = (done, dialect) => {
         const data = {CommitTimestamp: Spanner.COMMIT_TIMESTAMP};
@@ -1343,6 +1415,17 @@ describe('Spanner', () => {
   });
 
   describe('instanceConfigs', () => {
+    it('should have created the instance config', function (done) {
+      if (IS_EMULATOR_ENABLED) {
+        this.skip();
+      }
+      instanceConfig.get((err, metadata) => {
+        assert.ifError(err);
+        assert.strictEqual(metadata!.name, instanceConfig.formattedName_);
+        done();
+      });
+    });
+
     it('should list the available instanceConfigs', done => {
       spanner.getInstanceConfigs((err, instanceConfigs) => {
         assert.ifError(err);
@@ -1374,7 +1457,7 @@ describe('Spanner', () => {
         );
     });
 
-    it('should get an instanceConfig', function (done) {
+    it("should get an instanceConfig's metadata using getInstanceConfig", function (done) {
       if (IS_EMULATOR_ENABLED) {
         this.skip();
       }
@@ -1385,7 +1468,7 @@ describe('Spanner', () => {
       });
     });
 
-    it('should get an instanceConfig in promise mode', function (done) {
+    it("should get an instanceConfig's metadata in promise mode using getInstanceConfig", function (done) {
       if (IS_EMULATOR_ENABLED) {
         this.skip();
       }
@@ -1397,6 +1480,109 @@ describe('Spanner', () => {
           done();
         })
         .catch(done);
+    });
+
+    it("should get an instanceConfig's metadata using get", function (done) {
+      if (IS_EMULATOR_ENABLED) {
+        this.skip();
+      }
+      spanner
+        .instanceConfig('nam6')
+        .get()
+        .then(data => {
+          const instanceConfig = data[0];
+          assert(instanceConfig.displayName);
+          done();
+        })
+        .catch(done);
+    });
+
+    it("should list an instanceConfig's operations without filter", async function () {
+      if (IS_EMULATOR_ENABLED) {
+        this.skip();
+      }
+
+      const [operationsWithoutFilter] =
+        await spanner.getInstanceConfigOperations();
+      const operationForCurrentInstanceConfig = operationsWithoutFilter!.find(
+        operation =>
+          operation.name &&
+          operation.name.includes(instanceConfig.formattedName_)
+      );
+      assert.ok(operationForCurrentInstanceConfig);
+      assert.strictEqual(
+        operationForCurrentInstanceConfig!.metadata!.type_url,
+        'type.googleapis.com/google.spanner.admin.instance.v1.CreateInstanceConfigMetadata'
+      );
+    });
+
+    it("should list an instanceConfig's operations with filter", async function () {
+      if (IS_EMULATOR_ENABLED) {
+        this.skip();
+      }
+
+      const [operationsWithFilter] = await spanner.getInstanceConfigOperations({
+        filter: `(metadata.@type:type.googleapis.com/google.spanner.admin.instance.v1.CreateInstanceConfigMetadata) AND
+                    (name:${instanceConfig.formattedName_})`,
+      });
+      const operationForCurrentInstanceConfigWithFilter =
+        operationsWithFilter[0];
+      assert.ok(operationForCurrentInstanceConfigWithFilter);
+      const operationForCurrentInstanceConfigWithFilterMetadata =
+        CreateInstanceConfigMetadata.decode(
+          operationForCurrentInstanceConfigWithFilter!.metadata!
+            .value! as Uint8Array
+        );
+      assert.strictEqual(
+        operationForCurrentInstanceConfigWithFilterMetadata.instanceConfig!
+          .name,
+        `${instanceConfig.formattedName_}`
+      );
+    });
+
+    it('should update the instance config metadata', function (done) {
+      if (IS_EMULATOR_ENABLED) {
+        this.skip();
+      }
+      const newData = {
+        instanceConfig: {
+          displayName: 'new-display-name-' + shortUUID(),
+        },
+      };
+
+      instanceConfig.setMetadata(
+        newData,
+        execAfterOperationComplete(err => {
+          assert.ifError(err);
+
+          instanceConfig.get((err, metadata) => {
+            assert.ifError(err);
+            assert.strictEqual(
+              metadata!.displayName,
+              newData.instanceConfig.displayName
+            );
+            done();
+          });
+        })
+      );
+    });
+
+    it('should return true for instance config that exist', async function () {
+      if (IS_EMULATOR_ENABLED) {
+        this.skip();
+      }
+      const doesExist = await instanceConfig.exists();
+      assert.strictEqual(doesExist, true);
+    });
+
+    it('should return false for instance configs that do not exist', async function () {
+      if (IS_EMULATOR_ENABLED) {
+        this.skip();
+      }
+      const doesExist = await spanner
+        .instanceConfig('bad-instance-config')
+        .exists();
+      assert.strictEqual(doesExist, false);
     });
   });
 
@@ -6613,10 +6799,11 @@ describe('Spanner', () => {
           NumberValue: 0,
         };
 
-        beforeEach(() => {
-          return googleSqlTable.update(defaultRowValues).then(() => {
-            postgreSqlTable.update(defaultRowValues);
-          });
+        beforeEach(async () => {
+          await googleSqlTable.update(defaultRowValues);
+          if (!IS_EMULATOR_ENABLED) {
+            await postgreSqlTable.update(defaultRowValues);
+          }
         });
 
         const readConcurrentTransaction = (done, database, table) => {